Add GuissianiVersus2021.py
This is the initial commit for the GiussaniVersus python script that parses through Gaussian output files. This should be able to parse through Qchem files in the future and should have a cleaner output. It also may be preferable for the choosing of files to be done in a UI rather than having to manually type in the file names.